datacube: reorganize state and refactor initialization
