+++ .changeset/neat-chicken-pretend.md
@@ -1,0 +1,6 @@
+---
+'@finos/legend-application-query': patch
+'@finos/legend-application-repl': patch
+'@finos/legend-query-builder': patch
+'@finos/legend-data-cube': patch
+---
