--- packages/legend-application-repl/src/stores/LegendREPLDataCubeApplicationEngine.ts
@@ -1,154 +1,0 @@
-/**
- * Copyright (c) 2020-present, Goldman Sachs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import {
-  shouldDisplayVirtualAssistantDocumentationEntry,
-  APPLICATION_EVENT,
-} from '@finos/legend-application';
-import {
-  type DocumentationEntry,
-  formatDate,
-  LogEvent,
-  type PlainObject,
-} from '@finos/legend-shared';
-import {
-  DataCubeApplicationEngine,
-  type DataCubeSource,
-} from '@finos/legend-data-cube';
-import type { LegendREPLApplicationStore } from '../application/LegendREPLApplicationStore.js';
-import { LegendREPLDataCubeSource } from './LegendREPLDataCubeSource.js';
-
-export class LegendREPLDataCubeApplicationEngine extends DataCubeApplicationEngine {
-  private readonly application: LegendREPLApplicationStore;
-
-  constructor(application: LegendREPLApplicationStore) {
-    super();
-
-    this.application = application;
-  }
-
-  blockNavigation(
-    blockCheckers: (() => boolean)[],
-    onBlock?: ((onProceed: () => void) => void) | undefined,
-    onNativePlatformNavigationBlock?: (() => void) | undefined,
-  ) {
-    this.application.navigationService.navigator.blockNavigation(
-      blockCheckers,
-      onBlock,
-      onNativePlatformNavigationBlock,
-    );
-  }
-
-  unblockNavigation() {
-    this.application.navigationService.navigator.unblockNavigation();
-  }
-
-  getPersistedNumericSettingValue(key: string): number | undefined {
-    return this.application.settingService.getNumericValue(key);
-  }
-
-  getPersistedStringSettingValue(key: string): string | undefined {
-    return this.application.settingService.getStringValue(key);
-  }
-
-  getPersistedBooleanSettingValue(key: string): boolean | undefined {
-    return this.application.settingService.getBooleanValue(key);
-  }
-
-  getPersistedObjectSettingValue(key: string): object | undefined {
-    return this.application.settingService.getObjectValue(key);
-  }
-
-  persistSettingValue(
-    key: string,
-    value: string | number | boolean | object | undefined,
-  ): void {
-    this.application.settingService.persistValue(key, value);
-  }
-
-  // ---------------------------------- API ----------------------------------
-
-  getDocumentationURL(): string | undefined {
-    return this.application.documentationService.url;
-  }
-
-  getDocumentationEntry(key: string) {
-    return this.application.documentationService.getDocEntry(key);
-  }
-
-  shouldDisplayDocumentationEntry(entry: DocumentationEntry) {
-    return shouldDisplayVirtualAssistantDocumentationEntry(entry);
-  }
-
-  openLink(url: string) {
-    this.application.navigationService.navigator.visitAddress(url);
-  }
-
-  sendTelemetry(event: string, data: PlainObject) {
-    this.application.telemetryService.logEvent(event, data);
-  }
-
-  logDebug(message: string, ...data: unknown[]) {
-    this.application.logService.debug(
-      LogEvent.create(APPLICATION_EVENT.DEBUG),
-      message,
-      ...data,
-    );
-  }
-
-  debugProcess(processName: string, ...data: [string, unknown][]) {
-    this.application.logService.debug(
-      LogEvent.create(APPLICATION_EVENT.DEBUG),
-      `\n------ START DEBUG PROCESS: ${processName} ------`,
-      ...data.flatMap(([key, value]) => [`\n[${key.toUpperCase()}]:`, value]),
-      `\n------- END DEBUG PROCESS: ${processName} -------\n\n`,
-    );
-  }
-
-  logInfo(event: LogEvent, ...data: unknown[]) {
-    this.application.logService.info(event, ...data);
-  }
-
-  logWarning(event: LogEvent, ...data: unknown[]) {
-    this.application.logService.warn(event, ...data);
-  }
-
-  logError(event: LogEvent, ...data: unknown[]) {
-    this.application.logService.error(event, ...data);
-  }
-
-  logUnhandledError(error: Error) {
-    this.application.logUnhandledError(error);
-  }
-
-  logIllegalStateError(message: string, error?: Error) {
-    this.logError(
-      LogEvent.create(APPLICATION_EVENT.ILLEGAL_APPLICATION_STATE_OCCURRED),
-      message,
-      error,
-    );
-  }
-
-  // ---------------------------------- EVENT HANDLER ----------------------------------
-
-  override onNameChange(name: string, source: DataCubeSource) {
-    const timestamp =
-      source instanceof LegendREPLDataCubeSource ? source.timestamp : undefined;
-    this.application.layoutService.setWindowTitle(
-      `\u229E ${name}${timestamp ? ` - ${formatDate(new Date(timestamp), 'HH:mm:ss EEE MMM dd yyyy')}` : ''}`,
-    );
-  }
-}
