+++ packages/legend-data-cube/src/stores/view/grid/DataCubeGridClientEngine.ts
@@ -116,7 +116,7 @@ export const generateBackgroundColorUtilityClassName = (
 export const INTERNAL__GRID_CLIENT_DEFAULT_CACHE_BLOCK_SIZE = 500;
 export const INTERNAL__GRID_CLIENT_DEFAULT_ENABLE_PAGINATION = true;
 // NOTE: The cache block size is used by ag-grid to pre-allocate memory for the grid
-// so the value set must be reasonable, or else it can crash the application!
+// so the value set must be reasonable, or else it can crash the engine!
 export const INTERNAL__GRID_CLIENT_MAX_CACHE_BLOCK_SIZE = 1e4;
 
 export const INTERNAL__GRID_CLIENT_PIVOT_COLUMN_GROUP_COLOR_ROTATION_SIZE = 5;
@@ -287,7 +287,11 @@ async function getCastColumns(
 
   const lambda = new V1_Lambda();
   lambda.body.push(query);
-  const result = await view.engine.executeQuery(lambda, view.source);
+  const result = await view.engine.executeQuery(
+    lambda,
+    view.source,
+    view.dataCube,
+  );
 
   return result.result.builder.columns.map((column) => ({
     name: column.name,
@@ -360,7 +364,7 @@ export class DataCubeGridClientServerSideDataSource
             );
           } catch (error) {
             assertErrorThrown(error);
-            this.grid.view.application.alertError(error, {
+            this.grid.view.engine.alertError(error, {
               message: `Query Validation Failure: Can't retrieve pivot results column metadata. ${error.message}`,
             });
             // fail early since we can't proceed without the cast columns validated
@@ -413,10 +417,11 @@ export class DataCubeGridClientServerSideDataSource
       const result = await this.grid.view.engine.executeQuery(
         lambda,
         this.grid.view.source,
+        this.grid.view.dataCube,
       );
       const rowData = buildRowData(result.result.result, newSnapshot);
-      if (this.grid.view.engine.enableDebugMode) {
-        this.grid.view.application.debugProcess(
+      if (this.grid.view.dataCube.settings.enableDebugMode) {
+        this.grid.view.engine.debugProcess(
           `Execution`,
           ['Query', result.executedQuery],
           ['Config', `pagination=${this.grid.isPaginationEnabled}`],
@@ -461,10 +466,13 @@ export class DataCubeGridClientServerSideDataSource
         // When there are just too many rows (exceeding the maximum cache block size), we will fallback to a slightly less ideal
         // behavior by forcing a scroll top for every data fetch and also reset the cache block size to the default value to save memory
         if (rowData.length > INTERNAL__GRID_CLIENT_MAX_CACHE_BLOCK_SIZE) {
-          if (!this.grid.view.engine.gridClientSuppressLargeDatasetWarning) {
-            this.grid.view.application.alert({
+          if (
+            !this.grid.view.dataCube.settings
+              .gridClientSuppressLargeDatasetWarning
+          ) {
+            this.grid.view.engine.alert({
               message: `Large dataset (>${INTERNAL__GRID_CLIENT_MAX_CACHE_BLOCK_SIZE} rows) detected!`,
-              text: `Overall app performance can be impacted by large dataset due to longer query execution time and increased memory usage. At its limit, the application can crash!\nTo boost performance, consider enabling pagination while working with large dataset.`,
+              text: `Overall app performance can be impacted by large dataset due to longer query execution time and increased memory usage. At its limit, the engine can crash!\nTo boost performance, consider enabling pagination while working with large dataset.`,
               type: AlertType.WARNING,
               actions: [
                 {
@@ -476,7 +484,7 @@ export class DataCubeGridClientServerSideDataSource
                 {
                   label: 'Dismiss Warning',
                   handler: () => {
-                    this.grid.view.engine.setGridClientSuppressLargeDatasetWarning(
+                    this.grid.view.dataCube.settings.setGridClientSuppressLargeDatasetWarning(
                       true,
                     );
                   },
@@ -513,7 +521,7 @@ export class DataCubeGridClientServerSideDataSource
       }
     } catch (error) {
       assertErrorThrown(error);
-      this.grid.view.application.alertError(error, {
+      this.grid.view.engine.alertError(error, {
         message: `Data Fetch Failure: ${error.message}`,
       });
       params.fail();
@@ -525,7 +533,7 @@ export class DataCubeGridClientServerSideDataSource
   getRows(params: IServerSideGetRowsParams<unknown, unknown>) {
     this.fetchRows(params).catch((error: unknown) => {
       assertErrorThrown(error);
-      this.grid.view.application.logIllegalStateError(
+      this.grid.view.engine.logIllegalStateError(
         `Error ocurred while fetching data for grid should have been handled gracefully`,
         error,
       );
