+++ packages/legend-query-builder/src/components/data-cube/QueryBuilderDataCube.tsx
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-import { DataCube, DataCubeProvider } from '@finos/legend-data-cube';
+import { DataCube } from '@finos/legend-data-cube';
 import type { QueryBuilderState } from '../../stores/QueryBuilderState.js';
 import { observer } from 'mobx-react-lite';
-import { QueryBuilderDataCubeApplicationEngine } from '../../stores/data-cube/QueryBuilderDataCubeApplicationEngine.js';
 import {
   clsx,
   Dialog,
@@ -31,19 +30,12 @@ import { createDataCubeEngineFromQueryBuilder } from '../../stores/data-cube/Que
 const QueryBuilderDataCube = observer(
   (props: { queryBuilderState: QueryBuilderState }) => {
     const { queryBuilderState } = props;
-    const application = new QueryBuilderDataCubeApplicationEngine(
-      queryBuilderState.applicationStore,
-    );
     const engine = createDataCubeEngineFromQueryBuilder(queryBuilderState);
+
     if (!engine) {
       return null;
     }
-
-    return (
-      <DataCubeProvider application={application} engine={engine}>
-        <DataCube />
-      </DataCubeProvider>
-    );
+    return <DataCube engine={engine} />;
   },
 );
