+++ packages/legend-query-builder/src/index.ts
@@ -126,5 +126,4 @@ export * from './stores/filter/QueryBuilderFilterValueSpecificationBuilder.js';
 export * from './stores/QueryBuilderValueSpecificationHelper.js';
 export * from './stores/filter/QueryBuilderFilterState.js';
 export * from './stores/filter/QueryBuilderFilterStateBuilder.js';
-export * from './stores/data-cube/QueryBuilderDataCubeApplicationEngine.js';
 export * from './stores/data-cube/QueryBuilderDataCubeEngine.js';
